Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
pyotp==2.8.0
qrcode==7.4.2
pillow==10.0.1
twilio==8.10.0
flask-session==0.4.0
gunicorn==21.2.0
opencv-python==4.8.1.78
numpy==1.24.3
face-recognition==1.3.0